fix use of auth creds
